Projects/Non actual/ParentChildResearch/Version1/requirements.txt
python-docx>=0.8.11
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0


















